Optimization: move BFGS over to Optimization namespace
